requirements.txt
Django~=4.2.0
djangorestframework~=4.12.2
djangorestframework-simplejwt~=5.3.1
drf-spectacular>=0.27.2